requirements.txt
lxml
O365==2.1.8
msal==1.34.0
selenium
requests
tzlocal
python-dotenv
requests_oauthlib>=1.2.0

# Google Calendar API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

# MQTT
paho-mqtt==1.6.1

# REST API
flask==3.0.0
